Move user input to separate component UserInput
